temp

+++ src/utils/commands/saveSendLogs.ts
@@ -1,4 +1,4 @@
-import { DiscordEvent } from '../types/types';
+import { DiscordEvent, EventLog } from '../schema';
 import { assert } from '../assert';
 import fs from 'fs';
 import path from 'path';
@@ -6,12 +6,8 @@ import readline from 'readline';
 
 async function appendLinesToFile(filePath: string, lines: string[]): Promise<void> {
     try {
-        if (!fs.existsSync(filePath)) {
-            await fs.promises.writeFile(filePath, '', 'utf8');
-        }
-
+        if (!fs.existsSync(filePath)) await fs.promises.writeFile(filePath, '', 'utf8');
         const data = lines.join('\n') + '\n';
-
         await fs.promises.appendFile(filePath, data, 'utf8');
     } catch (error) {
         assert(error instanceof Error, 'Error should be an instance of Error');
@@ -29,41 +25,30 @@ async function readLastNLines(filePath: string, n: number): Promise<string[]> {
         crlfDelay: Infinity,
     });
 
+    // Read the file line by line, keeping only the last n lines
     const buffer: string[] = [];
-
-    // Read the file line by line
     for await (const line of rl) {
         buffer.push(line);
-        if (buffer.length > n) {
-            buffer.shift();
-        }
+        if (buffer.length > n) buffer.shift();
     }
 
     rl.close();
-
     return buffer;
 }
 
-type EventLog = {
-    subject: string;
-    summary: string;
-    calendarId: string;
-    messageId: string;
-    threadId: string;
-    guildId: string;
-};
-
-function formatDiscordEvent(event: DiscordEvent): string[] {
+function formatDiscordEvent(event: DiscordEvent, timestamp: Date): string[] {
     const { events, messageId, threadId, guildId } = event;
-    const eventLogs: EventLog[] = events.map((event) => {
+    const eventLogs = events.map((event) => {
         const { calendarId, subject, summary } = event;
-        return { subject, summary, calendarId, messageId, threadId, guildId };
-    });
+        return { timestamp, subject, summary, calendarId, messageId, threadId, guildId };
+    }) as EventLog[];
 
     return eventLogs.map((event) => JSON.stringify(event));
 }
+
 export async function saveSendLogs(successfulEvents: DiscordEvent[]) {
-    const logs: string[][] = successfulEvents.map(formatDiscordEvent);
+    const timestamp = new Date();
+    const logs: string[][] = successfulEvents.map((event) => formatDiscordEvent(event, timestamp));
     const logFilePath = path.join(__dirname, '../../data/sendLogs.txt');
     await appendLinesToFile(logFilePath, logs.flat());
 }
